+++ MysteryMud.Domain/Combat/Effect/Definitions/InstantDamageActionDefinition.cs
@@ -4,6 +4,6 @@ namespace MysteryMud.Domain.Combat.Effect.Definitions;
 
 public class InstantDamageActionDefinition : EffectActionDefinition
 {
-    public required Func<EffectContext, int> AmountFunc { get; init; }
+    public required Func<EffectContext, decimal> AmountFunc { get; init; }
     public required DamageKind Kind { get; init; }
 }
