+++ MysteryMud.Domain/Combat/Effect/Definitions/InstantHealActionDefinition.cs
@@ -2,5 +2,5 @@
 
 public class InstantHealActionDefinition : EffectActionDefinition
 {
-    public required Func<EffectContext, int> AmountFunc { get; init; }
+    public required Func<EffectContext, decimal> AmountFunc { get; init; }
 }
