+++ MysteryMud.GameData/Definitions/StatModifier.cs
@@ -6,5 +6,5 @@ public readonly struct StatModifier
 {
     public required StatKind Stat { get; init; }
     public required ModifierKind Modifier { get; init; }
-    public required int Value { get; init; }
+    public required decimal Value { get; init; }
 }
